# proto/gotocompany/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "gotocompany/stencil/v1beta1/stencil.proto",
    "version": "0.1.4"
  },
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1beta1/namespaces/{namespaceId}/schemas/{schemaId}": {
      "delete": {
        "summary": "Delete specified schema",
        "operationId": "StencilService_DeleteSchema",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1beta1DeleteSchemaResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "schemaId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "schema"
        ]
      },
      "patch": {
        "summary": "Update only schema metadata",
        "operationId": "StencilService_UpdateSchemaMetadata",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1beta1UpdateSchemaMetadataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "schemaId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "compatibility": {
                  "$ref": "#/definitions/SchemaCompatibility"
                }
              }
            }
          }
        ],
        "tags": [
          "schema"
        ]
      }
    },
...
    "/v1beta1/namespaces/{namespaceId}/schemas/{schemaId}/versions/{versionId}": {
      "delete": {
        "summary": "Delete specified version of the schema",
        "operationId": "StencilService_DeleteVersion",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1beta1DeleteVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "schemaId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "schema",
          "version"
        ]
      }
    },
    "/v1beta1/namespaces/{namespace_id}/schemas/{schema_id}/types/{type_name}/lineage": {
      "get": {
        "summary": "Get lineage graph for a schema",
        "operationId": "StencilService_GetLineage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1beta1LineageResponse"
            }
          },
          "400": {
            "description": "Invalid request parameters (for example unsupported direction).",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "schema_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Schema container ID (for example, esb-log-entities)."
          },
          {
            "name": "type_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Root proto message for lineage traversal. Accepts fully-qualified names (for example, gojek.esb.types.Location)."
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32",
            "description": "Maximum traversal depth. Defaults to 10 when omitted or <= 0."
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "both",
              "downstream",
              "upstream"
            ],
            "default": "both",
            "description": "Traversal direction for lineage discovery."
          }
        ],
        "tags": [
          "schema"
        ]
      }
    },
...
  },
  "definitions": {
    "SchemaCompatibility": {
      "type": "string",
      "enum": [
        "COMPATIBILITY_UNSPECIFIED",
        "COMPATIBILITY_BACKWARD",
        "COMPATIBILITY_BACKWARD_TRANSITIVE",
        "COMPATIBILITY_FORWARD",
        "COMPATIBILITY_FORWARD_TRANSITIVE",
        "COMPATIBILITY_FULL",
        "COMPATIBILITY_FULL_TRANSITIVE"
      ],
      "default": "COMPATIBILITY_UNSPECIFIED"
    },
    "SchemaFormat": {
      "type": "string",
      "enum": [
        "FORMAT_UNSPECIFIED",
        "FORMAT_PROTOBUF",
        "FORMAT_AVRO",
        "FORMAT_JSON"
      ],
      "default": "FORMAT_UNSPECIFIED"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1beta1DeleteSchemaResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1beta1DeleteVersionResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "v1beta1LineageResponse": {
      "type": "object",
      "properties": {
        "root_schema": {
          "$ref": "#/definitions/v1beta1RootSchemaRef"
        },
        "direction": {
          "type": "string",
          "enum": [
            "both",
            "downstream",
            "upstream"
          ]
        },
        "downstream": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1beta1LineageSchema"
          }
        },
        "upstream": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1beta1LineageSchema"
          }
        },
        "summary": {
          "$ref": "#/definitions/v1beta1LineageSummary"
        }
      }
    },
    "v1beta1LineageSchema": {
      "type": "object",
      "properties": {
        "namespace_id": {
          "type": "string"
        },
        "schema_id": {
          "type": "string"
        },
        "type_name": {
          "type": "string"
        },
        "level": {
          "type": "integer",
          "format": "int32"
        },
        "path": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1beta1LineageSummary": {
      "type": "object",
      "properties": {
        "downstream_count": {
          "type": "integer",
          "format": "int32"
        },
        "upstream_count": {
          "type": "integer",
          "format": "int32"
        },
        "total_count": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v1beta1RootSchemaRef": {
      "type": "object",
      "properties": {
        "namespace_id": {
          "type": "string"
        },
        "schema_id": {
          "type": "string"
        },
        "type_name": {
          "type": "string"
        }
      }
    },
...
    "v1beta1UpdateSchemaMetadataResponse": {
      "type": "object",
      "properties": {
        "format": {
          "$ref": "#/definitions/SchemaFormat"
        },
        "compatibility": {
          "$ref": "#/definitions/SchemaCompatibility"
        },
        "authority": {
          "type": "string"
        }
      }
    }
  }
}